Move get_expected_length to pixel_data_handlers.util
